Refactor elements to functional Components
React elements doesn't take functional Components as a child so to
render the NameComponent and HandleComponent, we must either use another
React.createElement() method or call the function Component itself.

Line 38 demonstrates using React.createElement().
Line 43 demonstrates calling the functional HandleComponent.

React.createElement() takes the NameComponent as the first argument and
any props that need to passed in as the second argument. The props must
be named and match the prop name in the Component itself.
The props must be an object of key/value pair as seen on line 39-41.

HandleComponent() is called on line 43 and the prop name is also an
object of key/value pair that's passed through in order to access that
specific data.